statistics chart display config finished, date range selecting finished

+++ android/Petimo/app/src/main/java/de/tud/nhd/petimo/model/db/MonitorDay.java
@@ -170,7 +170,7 @@ public class MonitorDay {
      * Calculate the total duration of all monitored blocks of this day
      * @return the total duration
      */
-    public int getDuration(){
+    public long getDuration(){
         int duration = 0;
         if (!monitorBlocks.isEmpty())
             for (MonitorBlock block : monitorBlocks)
